+++ src/Sulu/Bundle/ContentBundle/Tests/Unit/Content/Structure/ExcerptStructureExtensionTest.php
@@ -16,6 +16,7 @@ use Sulu\Bundle\SearchBundle\Search\Factory;
 use Sulu\Component\Content\Compat\StructureInterface;
 use Sulu\Component\Content\Compat\StructureManagerInterface;
 use Sulu\Component\Content\ContentTypeManagerInterface;
+use Sulu\Component\Content\Export\ContentExportManagerInterface;
 
 class ExcerptStructureExtensionTest extends \PHPUnit_Framework_TestCase
 {
@@ -35,11 +36,13 @@ class ExcerptStructureExtensionTest extends \PHPUnit_Framework_TestCase
         $contentTypeManager = $this->prophesize(ContentTypeManagerInterface::class);
         $factory = $this->prophesize(Factory::class);
         $node = $this->prophesize(NodeInterface::class);
+        $contentExportManager = $this->prophesize(ContentExportManagerInterface::class);
 
         $excerptExtension = new ExcerptStructureExtension(
             $structureManager->reveal(),
             $contentTypeManager->reveal(),
-            $factory->reveal()
+            $factory->reveal(),
+            $contentExportManager->reveal()
         );
 
         $excerptExtension->save($node->reveal(), [], 'sulu_io', 'de');
@@ -61,11 +64,13 @@ class ExcerptStructureExtensionTest extends \PHPUnit_Framework_TestCase
         $contentTypeManager = $this->prophesize(ContentTypeManagerInterface::class);
         $factory = $this->prophesize(Factory::class);
         $node = $this->prophesize(NodeInterface::class);
+        $contentExportManager = $this->prophesize(ContentExportManagerInterface::class);
 
         $excerptExtension = new ExcerptStructureExtension(
             $structureManager->reveal(),
             $contentTypeManager->reveal(),
-            $factory->reveal()
+            $factory->reveal(),
+            $contentExportManager->reveal()
         );
 
         $excerptExtension->load($node->reveal(), 'sulu_io', 'de');